# server/openapi/v2.json
{
  "components": {
    "schemas": {
...
      "AppItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modified_at": {
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "enabled",
          "version",
          "created_at",
          "modified_at"
        ],
        "title": "AppItem",
        "type": "object"
      },
      "AppItemResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AppItem"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "AppItemResponse",
        "type": "object"
      },
...
      "AppKeyItemResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AppKeyItem"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "AppKeyItemResponse",
        "type": "object"
      },
      "AppKeyList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AppKeyItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "AppKeyList",
        "type": "object"
      },
      "AppKeyListResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AppKeyList"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "AppKeyListResponse",
        "type": "object"
      },
...
      "AppList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AppListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "max_page": {
            "title": "Max Page",
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "max_page",
          "items"
        ],
        "title": "AppList",
        "type": "object"
      },
...
      "AppListResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AppList"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "AppListResponse",
        "type": "object"
      },
...
      "AppProviderResultResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AppProviderResult"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "AppProviderResultResponse",
        "type": "object"
      },
...
      "BulkUpdateGatewayParams": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "gateways": {
            "items": {
              "$ref": "#/components/schemas/GatewayBulkTargetParams"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Gateways",
            "type": "array"
          }
        },
        "required": [
          "enabled",
          "gateways"
        ],
        "title": "BulkUpdateGatewayParams",
        "type": "object"
      },
...
      "BulkUpdateGatewayResultResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BulkUpdateGatewayResult"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "BulkUpdateGatewayResultResponse",
        "type": "object"
      },
...
      "CreateAppParams": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateAppParams",
        "type": "object"
      },
...
      "CreatedApp": {
        "properties": {
          "api_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "api_key_id": {
            "title": "Api Key Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modified_at": {
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "enabled",
          "version",
          "created_at",
          "modified_at",
          "api_key_id",
          "api_key"
        ],
        "title": "CreatedApp",
        "type": "object"
      },
...
      "CreatedAppKeyResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreatedAppKey"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "CreatedAppKeyResponse",
        "type": "object"
      },
      "CreatedAppResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreatedApp"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "CreatedAppResponse",
        "type": "object"
      },
...
      "DeletedApp": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "deleted"
        ],
        "title": "DeletedApp",
        "type": "object"
      },
      "DeletedAppResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeletedApp"
          },
          "msg": {
            "default": "ok",
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "DeletedAppResponse",
        "type": "object"
      },
...
      "ErrorResponseModel": {
        "properties": {
          "code": {
            "default": "internal.error",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "success": {
            "default": false,
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "msg"
        ],
        "title": "ErrorResponseModel",
        "type": "object"
      },
...
      "SetAppProviderParams": {
        "additionalProperties": false,
        "properties": {
          "provider_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Provider Id",
            "type": "string"
          }
        },
        "required": [
          "provider_id"
        ],
        "title": "SetAppProviderParams",
        "type": "object"
      },
...
      "UpdateAppParams": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "UpdateAppParams",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "PersonalAccessToken": {
        "bearerFormat": "PAT",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "rpc_gateway_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "rpc-gateway-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/apps": {
      "get": {
        "operationId": "apps-list_apps",
        "parameters": [
          {
            "in": "query",
            "name": "start_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start At"
            }
          },
          {
            "in": "query",
            "name": "end_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End At"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 320,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "List Apps",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "apps:read"
        ]
      },
      "post": {
        "description": "Create an App, its first API Key, and its catalog snapshot of Gateways.",
        "operationId": "apps-create_app",
        "parameters": [
          {
            "description": "Replays a successful PAT write with the same account, key, and request for 24 hours.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedAppResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "Create App",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "app-keys:write",
          "apps:write"
        ]
      }
    },
    "/v2/apps/{app_id}": {
      "delete": {
        "operationId": "apps-delete_app",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "description": "Replays a successful PAT write with the same account, key, and request for 24 hours.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedAppResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "Delete App",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "apps:write"
        ]
      },
      "get": {
        "operationId": "apps-get_app",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "Get App",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "apps:read"
        ]
      },
      "patch": {
        "operationId": "apps-update_app",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "description": "Replays a successful PAT write with the same account, key, and request for 24 hours.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "Update App",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "apps:write"
        ]
      }
    },
    "/v2/apps/{app_id}/api-keys": {
      "get": {
        "operationId": "apps-list_api_keys",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppKeyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "List Api Keys",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "app-keys:read"
        ]
      }
    },
    "/v2/apps/{app_id}/api-keys/rotate": {
      "post": {
        "description": "Create an API Key and move the prior active Key into its configured grace period.",
        "operationId": "apps-rotate_api_key",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "description": "Replays a successful PAT write with the same account, key, and request for 24 hours.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedAppKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "Rotate Api Key",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "app-keys:write"
        ]
      }
    },
    "/v2/apps/{app_id}/api-keys/{key_id}/revoke": {
      "post": {
        "operationId": "apps-revoke_api_key",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "description": "Replays a successful PAT write with the same account, key, and request for 24 hours.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppKeyItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "Revoke Api Key",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "app-keys:write"
        ]
      }
    },
    "/v2/apps/{app_id}/gateways": {
      "patch": {
        "description": "Atomically set enabled for Gateways belonging to one App using optimistic versions.",
        "operationId": "apps-bulk_update_app_gateways",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          },
          {
            "description": "Replays a successful PAT write with the same account, key, and request for 24 hours.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateGatewayParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateGatewayResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "PersonalAccessToken": []
          }
        ],
        "summary": "Bulk Update App Gateways",
        "tags": [
          "apps"
        ],
        "x-required-pat-scopes": [
          "gateways:write"
        ]
      }
    },
    "/v2/apps/{app_id}/provider": {
      "delete": {
        "description": "Remove the Provider association and its automatically managed Route Targets.",
        "operationId": "apps-clear_app_provider",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppProviderResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Clear App Provider",
        "tags": [
          "apps"
        ]
      },
      "put": {
        "description": "Associate an enabled account Provider and publish its matching Endpoints to this App.",
        "operationId": "apps-set_app_provider",
        "parameters": [
          {
            "in": "path",
            "name": "app_id",
            "required": true,
            "schema": {
              "title": "App Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAppProviderParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppProviderResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Set App Provider",
        "tags": [
          "apps"
        ]
      }
    },
...
  }
}